Update cosmetic_mfds.py
Added several functions to handle graph pairs (that is, graph manifolds where each piece is a SFS over a disk). We now have a rudimentary test to certify that such a manifold is chiral, and a rudimentary test to certify that two such are distinct. Neither test is exhaustive (meaning, it misses some True answers) but can be trusted when it returns True.