Update cosmetic_mfds.py
Lots of updates, mainly to find_common_fillings and the routines that feed into it.

At this time, find_common_fillings seems to work well at finding common fillings of M and N that are systole-short on M and kinda-low-volume on N.

For full correctness, the same code needs to be run with M and N reversed. We may choose to incorporate this automatically.