Updated Alexander_tests, prune_using_invariants
Updated the Alexander_tests function to return the Alexander polynomial itself, in addition to the other data.

Updated the prune_using_invariants function to implement the Daemi-Eismeier-Lidman test on nontriviality of the Alexander polynomial. By default, this test is turned off (Alex=False) because their preprint is not out yet. When the preprint appears, we can turn it on by default.