Added verified systoles; made HFK use more systematic
Added the use of verified_systole_with_drilling, to compute verified systoles.

Also, reduced the number of flag variables passed to the main computation functions for knots. The flags slope_method and use_NiWu have been disabled. All that remains is use_HFK. If true, we use all the HFK theorems, including the work of Hansselman and DLME. If false, we use none of that stuff.